+++ src/helper/configs/aruntime/project/saleor.go
@@ -1,0 +1,13 @@
+package project
+
+func init() {
+	RegisterDockerConfGenerator("saleor", MakeConfSaleor)
+}
+
+func MakeConfSaleor(projectName string) {
+	MakeDockerfile(projectName, "python/Dockerfile", "python.Dockerfile")
+	MakeDBDockerfile(projectName)
+	MakeRedisDockerfile(projectName)
+	MakeScriptsConf(projectName)
+	MakeClaudeDockerfile(projectName)
+}